Move files related to crypto
